Second Calculation and Continued Reporting
This includes a prompt inquiring how much a 12 pack of beer costs for
the user and captures that value into a variable.

A calculation is used to determine how much the user spends a year on
bottles of beer.

Reports the user’s yearly cost on beer bottles via an alert and also
reports it to the console. The amount of the cost has been fixed in
order to only include up to 2 decimal places.